Fix Mem_Calloc and add the public interface to it (mem.h)

+++ goldsrc/engine/fixme/mem.cpp
@@ -5,5 +5,5 @@
 
 void *Mem_Calloc(int number, size_t size)
 {
-	return calloc((int)number, size);
+	return calloc((size_t)number, size);
 };
